Initial state handling (#25)
This PR tackles #23 , we simplified the code:

1. We always pass a None if we want to rely on the `init_state_fn`,
simplifying the code. This is also similar how an initial state is
provided in LSTMS in PyTorch. If a user is interested in getting the
initial state where we passed None, the initial_state_fn can still be
called.
2. We further have now an explicit initial_state_fn that puts zeroes
everywhere. This allows users to understand what such a function could
look like! Even though we now set the Acados solver state two times to
zero....
3. We deleted some deprecated MPC class code!

---------

Co-authored-by: Confectio <leonard.fichtner@gmail.com>