Make Parallelization more accessible (#52)
The number of threads to use for parallelizing the mpc batch
calculations previously had to be set in the ocp formulation.
It can now be set dynamically at any point in time, which is exposed in
the Mpc class by the property `num_threads_in_batch_methods` and the
equally named constructor argument.

Further, it was actually not set at all by default, despite large
computational speedups being enabled by it.
Hence, there is now a new field `num_threads_mpc` in
the SacFopAlgorithmConfig (SacFop is currently the only architecture
using mpc batch calculations).
It is set to  `4` threads by default.

Enabled by:
[acados/acados/#1483](https://github.com/acados/acados/pull/1483)