Standardize example structure (#102)
This PR tries to clean up the structure of the example folders:
- Configs are directly at the place where they are required. Envs and
Controllers get seperate cfgs.
- The factory code for the acados_ocp is seperated in its own module.
- Different param interfaces are described by literals.

Every parameter that is easily switchable between learnable and not
learnable becomes an actual ``Parameter`` instance. Everything else is
put into the configs.

The configurations classes expose some customization but they are not
exhaustive (e.g., acados solver options are not generally exposed, or
slack variable weights, etc.), because that would essentially rebuild at
least the acados interface. The guideline is that everytime someone
wants to configure something, they can then expose it in the config.

closes #96