Improving reproducibility (#132)
This PR attempts to enforce reproducibility during validation and
training.

Essentially, the proposed approach consists in 1) creating a
`np.random.Genrator` with the specified run seed and assigning it to the
trainer, and 2) using said `np.random.Genrator` to generate a
(hopefully, kinda unique) seed every time `env.reset` is called.