Add Beta Distribution as an optional BoundedDistribution (#147)
Previously, sampling from a bounded space was always done via a
TanhNormal .
This PR implements an alternative distribution, the (unimodal) Beta Distribution,
and interfaces it to the SAC algorithm variants.
Also comes with a test that implements sanity checks for this new
distribution,
and manually tested for the run scripts of SAC-ZOP and SAC-FOP on Cart
Pole.