Fix hvac default param out of param space (#187)
The forecast-related parameter bound in the hvac example were too tight
and lead to `default_param` to be out of `controller.param_space` when
using historical data.

This PR fixes that by setting high values for `price` and `solar`. 

Note: A more systematic way would be to extrapolate these bounds based
on historical data.

---------

Co-authored-by: Jasper Hoffmann <hoffmaja@informatik.uni-freiburg.de>